requirements.txt
beautifulsoup4==4.12.2
markdownify==0.11.6
oauthlib==3.2.2
openai==1.2.2
readability-lxml==0.8.1
requests==2.31.0
requests-oauthlib==1.3.1
selenium==4.15.2
soupsieve==2.5
